split RawSearchResponse code
